+++ src/effect/elliptical-mask.ts
@@ -1,7 +1,8 @@
 import { Movie } from '../movie'
 import { Dynamic, val } from '../util'
 import { Visual } from './visual'
-import { Visual as VisualLayer } from '../layer/index'
+import { Visual2D, VisualBase as VisualBaseLayer } from '../layer/index'
+import { get2DRenderingContext } from '../compatibility-utils'
 
 export class EllipticalMaskOptions {
   x: Dynamic<number>
@@ -41,14 +42,15 @@ export class EllipticalMask extends Visual {
     this.startAngle = options.startAngle || 0
     this.endAngle = options.endAngle !== undefined ? options.endAngle : 2 * Math.PI
     this.anticlockwise = options.anticlockwise || false
-    // for saving image data before clearing
+
+    // For buffering if not using a view
     this._tmpCanvas = document.createElement('canvas')
     this._tmpCtx = this._tmpCanvas.getContext('2d')
   }
 
-  apply (target: Movie | VisualLayer, reltime: number): void {
-    const ctx = target.cctx
-    const canvas = target.canvas
+  apply (target: Movie | VisualBaseLayer, reltime: number): void {
+    const ctx = get2DRenderingContext(target)
+
     const x = val(this, 'x', reltime)
     const y = val(this, 'y', reltime)
     const radiusX = val(this, 'radiusX', reltime)
@@ -57,11 +59,16 @@ export class EllipticalMask extends Visual {
     const startAngle = val(this, 'startAngle', reltime)
     const endAngle = val(this, 'endAngle', reltime)
     const anticlockwise = val(this, 'anticlockwise', reltime)
+
+    if (target instanceof VisualBaseLayer && !(target instanceof Visual2D))
+      throw new Error('EllipticalMask effect applied to a non-2D layer that does not have a view!')
+
     this._tmpCanvas.width = target.canvas.width
     this._tmpCanvas.height = target.canvas.height
-    this._tmpCtx.drawImage(canvas, 0, 0)
+    this._tmpCtx.drawImage(target.canvas, 0, 0)
+    const output = this._tmpCanvas
 
-    ctx.clearRect(0, 0, canvas.width, canvas.height)
+    ctx.clearRect(0, 0, output.width, output.height)
     ctx.save() // idk how to preserve clipping state without save/restore
     // create elliptical path and clip
     ctx.beginPath()
@@ -69,7 +76,7 @@ export class EllipticalMask extends Visual {
     ctx.closePath()
     ctx.clip()
     // render image with clipping state
-    ctx.drawImage(this._tmpCanvas, 0, 0)
+    ctx.drawImage(output, 0, 0)
     ctx.restore()
   }
 }