+++ src/layer/image.ts
@@ -1,9 +1,9 @@
-import { Visual } from './visual'
+import { Visual2D } from './visual-2d'
 import { VisualSourceMixin, VisualSourceOptions } from './visual-source'
 
 type ImageOptions = VisualSourceOptions
 
-class Image extends VisualSourceMixin(Visual) {
+class Image extends VisualSourceMixin(Visual2D) {
   constructor (options: ImageOptions) {
     if (typeof (options.source) === 'string') {
       const img = document.createElement('img')